refactor: separate sum result and diff result
