+++ scan/scanner.go
@@ -1,7 +1,7 @@
 package scan
 
 type Scanner interface {
-	Scan() (Result, error)
+	Scan() (SumResult, error)
 	Diff(base Scanner) (DiffResult, error)
 	Export(outputType, filename string) error
 	ExportDiff(base Scanner, outputType, filename string) error
